Topic_scoring/study_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==6.0.1
django-cors-headers==4.7.0
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.10
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
gunicorn==23.0.0
idna==3.10
pdfplumber==0.11.5
psycopg2-binary==2.9.10
pypdf==5.3.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.5
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
